requirements.txt
flask==3.0.3
requests==2.32.3
python-dotenv==1.0.1
pydub==0.25.1
gTTS==2.5.1
elevenlabs==1.2.0
ffmpeg-python==0.2.0
werkzeug==3.0.3
PyPDF2==3.0.1
python-docx==1.1.0